move taskbar related components into separated directory
