+++ src/components/TaskBar/Divider.styles.tsx
