+++ src/components/TaskBar/Divider.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { Root } from "@components/TaskBar/Divider.styles";
+
+export interface TaskBarDividerProps extends React.HTMLAttributes<HTMLDivElement> {}
+
+export function TaskBarDivider(props: TaskBarDividerProps) {
+    return <Root {...props} />;
+}
